% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DromaSet-class.R
\name{availableMolecularProfiles}
\alias{availableMolecularProfiles}
\alias{availableMolecularProfiles,DromaSet-method}
\title{Get Available Molecular Profile Types}
\usage{
availableMolecularProfiles(object, include_db = TRUE)

\S4method{availableMolecularProfiles}{DromaSet}(object, include_db = TRUE)
}
\arguments{
\item{object}{A DromaSet object}

\item{include_db}{Logical, whether to include profiles available in the database but not loaded (default: TRUE)}
}
\value{
Character vector of available molecular profile types
}
\description{
Returns the types of molecular profiles available for a DromaSet
}
