% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiDromaSet-class.R
\name{availableProjects}
\alias{availableProjects}
\alias{availableProjects,MultiDromaSet-method}
\title{Get Available Projects in MultiDromaSet}
\usage{
availableProjects(object)

\S4method{availableProjects}{MultiDromaSet}(object)
}
\arguments{
\item{object}{A MultiDromaSet object}
}
\value{
Character vector of project names
}
\description{
Returns the names of projects available in a MultiDromaSet
}
