% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DromaSet-class.R
\name{availableTreatmentResponses}
\alias{availableTreatmentResponses}
\alias{availableTreatmentResponses,DromaSet-method}
\title{Get Available Treatment Response Types}
\usage{
availableTreatmentResponses(object, include_db = TRUE)

\S4method{availableTreatmentResponses}{DromaSet}(object, include_db = TRUE)
}
\arguments{
\item{object}{A DromaSet object}

\item{include_db}{Logical, whether to include data available in the database but not loaded (default: TRUE)}
}
\value{
Character vector of available treatment response types
}
\description{
Returns the types of treatment response data available for a DromaSet
}
