% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncMatrixDatabase.R
\name{closeCTRDatabase}
\alias{closeCTRDatabase}
\title{Close CTRDB Database Connection}
\usage{
closeCTRDatabase(connection = NULL)
}
\arguments{
\item{connection}{Optional database connection object. If NULL, uses global connection}
}
\value{
TRUE if successfully disconnected
}
\description{
Closes the connection to the CTRDB database
}
\examples{
\dontrun{
# Close CTRDB database connection
closeCTRDatabase()
}
}
