% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DROMA_SQLManager.R
\name{closeDROMADatabase}
\alias{closeDROMADatabase}
\title{Close DROMA Database Connection}
\usage{
closeDROMADatabase(connection = NULL)
}
\arguments{
\item{connection}{Optional database connection object. If NULL, uses global connection.}
}
\value{
TRUE if successfully disconnected
}
\description{
Closes the connection to the DROMA database
}
