% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTRDB_SQLManager.R
\name{connectCTRDatabase}
\alias{connectCTRDatabase}
\title{Connect to CTRDB Database}
\usage{
connectCTRDatabase(db_path = file.path(path.expand("sql_db"), "ctrdb.sqlite"))
}
\arguments{
\item{db_path}{Path to the CTRDB SQLite database file}
}
\value{
A database connection object
}
\description{
Establishes a connection to the Clinical Trial Response Database (CTRDB) SQLite database
}
\examples{
\dontrun{
# Connect to CTRDB database
con <- connectCTRDatabase("path/to/ctrdb.sqlite")
}
}
