% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DROMA_SQLManager.R
\name{connectDROMADatabase}
\alias{connectDROMADatabase}
\title{Connect to DROMA Database}
\usage{
connectDROMADatabase(
  db_path = file.path(path.expand("sql_db"), "droma.sqlite")
)
}
\arguments{
\item{db_path}{Path to the SQLite database file}
}
\value{
A database connection object
}
\description{
Establishes a connection to the DROMA SQLite database
}
