% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DromaSet-class.R
\docType{class}
\name{DromaSet-class}
\alias{DromaSet-class}
\title{DromaSet Class}
\description{
A class to represent a DROMA project dataset with drug response and omics data.
Each DromaSet contains treatment response data (drug sensitivity) and multiple molecular profiles
(omics data) for a specific project.
}
\section{Slots}{

\describe{
\item{\code{name}}{Character string, the name of the dataset (e.g., "gCSI", "CCLE")}

\item{\code{treatmentResponse}}{List containing drug response data matrices}

\item{\code{molecularProfiles}}{List containing different types of molecular data (omics)}

\item{\code{sampleMetadata}}{Data frame with sample annotations}

\item{\code{treatmentMetadata}}{Data frame with drug annotations}

\item{\code{datasetType}}{Character string, the type of dataset (e.g., "CellLine", "PDX", "PDO")}

\item{\code{db_info}}{List containing database connection information}
}}

