% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DromaSet-class.R
\name{DromaSet}
\alias{DromaSet}
\title{Create a DromaSet Object}
\usage{
DromaSet(
  name,
  treatmentResponse = list(),
  molecularProfiles = list(),
  sampleMetadata = data.frame(),
  treatmentMetadata = data.frame(),
  datasetType = NA_character_,
  db_info = list()
)
}
\arguments{
\item{name}{Character string, the name of the dataset (e.g., "gCSI", "CCLE")}

\item{treatmentResponse}{List containing drug response data matrices}

\item{molecularProfiles}{List containing different types of molecular data (omics)}

\item{sampleMetadata}{Data frame with sample annotations}

\item{treatmentMetadata}{Data frame with drug annotations}

\item{datasetType}{Character string, the type of dataset (e.g., "CellLine", "PDX", "PDO")}

\item{db_info}{List containing database connection information}
}
\value{
A DromaSet object
}
\description{
Creates a DromaSet object to store drug response and omics data for a specific project
}
\examples{
\dontrun{
# Create a DromaSet object for gCSI data
gCSI <- DromaSet(name = "gCSI",
                 datasetType = "CellLine",
                 db_info = list(db_path = "~/droma.sqlite", db_group = "gCSI"))
}
}
