% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiDromaSet-class.R
\name{getDromaSet}
\alias{getDromaSet}
\alias{getDromaSet,MultiDromaSet-method}
\title{Get DromaSet by Project Name}
\usage{
getDromaSet(object, project_name)

\S4method{getDromaSet}{MultiDromaSet}(object, project_name)
}
\arguments{
\item{object}{A MultiDromaSet object}

\item{project_name}{Character, the name of the project to retrieve}
}
\value{
A DromaSet object
}
\description{
Retrieves a specific DromaSet from a MultiDromaSet by project name
}
