% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiDromaSet-class.R
\name{getOverlappingSamples}
\alias{getOverlappingSamples}
\alias{getOverlappingSamples,MultiDromaSet-method}
\title{Get Overlapping Samples Between Projects}
\usage{
getOverlappingSamples(object, projects = NULL)

\S4method{getOverlappingSamples}{MultiDromaSet}(object, projects = NULL)
}
\arguments{
\item{object}{A MultiDromaSet object}

\item{projects}{Character vector, specific projects to check for overlap (default: all projects)}
}
\value{
A list containing overlapping sample information
}
\description{
Finds samples that are present in multiple projects within a MultiDromaSet
}
