% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DROMA_SQLManager.R
\name{listDROMAProjects}
\alias{listDROMAProjects}
\title{List Available Projects in DROMA Database}
\usage{
listDROMAProjects(
  connection = NULL,
  show_names_only = FALSE,
  project_data_types = NULL
)
}
\arguments{
\item{connection}{Optional database connection object. If NULL, uses global connection}

\item{show_names_only}{Logical, if TRUE returns only a character vector of project names}

\item{project_data_types}{Character, project name to get specific data types for}
}
\value{
A data frame with project information or a character vector of project names or data types
}
\description{
Lists all projects available in the DROMA database
}
