% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiDromaSet-class.R
\name{mergeSampleMetadata}
\alias{mergeSampleMetadata}
\title{Helper function to merge sample metadata from multiple DromaSets}
\usage{
mergeSampleMetadata(DromaSets)
}
\arguments{
\item{DromaSets}{List of DromaSet objects}
}
\value{
Merged data frame of sample metadata
}
\description{
Helper function to merge sample metadata from multiple DromaSets
}
