% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiDromaSet-class.R
\name{mergeTreatmentMetadata}
\alias{mergeTreatmentMetadata}
\title{Helper function to merge treatment metadata from multiple DromaSets}
\usage{
mergeTreatmentMetadata(DromaSets)
}
\arguments{
\item{DromaSets}{List of DromaSet objects}
}
\value{
Merged data frame of treatment metadata
}
\description{
Helper function to merge treatment metadata from multiple DromaSets
}
