% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiDromaSet-database.R
\name{removeDromaSetFromMulti}
\alias{removeDromaSetFromMulti}
\title{Remove DromaSet from MultiDromaSet}
\usage{
removeDromaSetFromMulti(multi_set, projects, update_metadata = TRUE)
}
\arguments{
\item{multi_set}{A MultiDromaSet object}

\item{projects}{Character, name of the project to remove}

\item{update_metadata}{Logical, whether to update merged metadata (default: TRUE)}
}
\value{
Updated MultiDromaSet object
}
\description{
Removes a DromaSet object from an existing MultiDromaSet
}
\examples{
\dontrun{
# Remove a project from MultiDromaSet
multi_set <- removeDromaSetFromMulti(multi_set, "CCLE")
}
}
