% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DromaSet-class.R
\name{show,DromaSet-method}
\alias{show,DromaSet-method}
\title{Show Method for DromaSet objects}
\usage{
\S4method{show}{DromaSet}(object)
}
\arguments{
\item{object}{A DromaSet object}
}
\value{
NULL, prints information to console
}
\description{
Displays information about a DromaSet object
}
