% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiDromaSet-class.R
\name{show,MultiDromaSet-method}
\alias{show,MultiDromaSet-method}
\title{Show Method for MultiDromaSet objects}
\usage{
\S4method{show}{MultiDromaSet}(object)
}
\arguments{
\item{object}{A MultiDromaSet object}
}
\value{
NULL, prints information to console
}
\description{
Displays information about a MultiDromaSet object
}
