% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DromaSet-class.R
\name{zscoreNormalize}
\alias{zscoreNormalize}
\title{Z-score normalization for omics data}
\usage{
zscoreNormalize(mat)
}
\arguments{
\item{mat}{A matrix or data frame with genes in rows and samples in columns}
}
\value{
A normalized matrix with Z-scores
}
\description{
Apply Z-score normalization to omics data at the gene level
}
